Move `test_patients_age_on`
The `patients` table moved from `tpp` to `core`, so its associated tests
should also move from `test_tpp` to `test_core`.
